Stricter unicode highlighting in vscode
I use <option>+['ao] for æøå in macOS and sometimes I accidentally
insert a non-breaking space right after due to pressing <option>+space.

Invisible unicode characters are not highlighted by default in vscode if
surrounded by only non-latin looking characters, e.g. in "for å gi et
eksempel" a non-breaking space after 'å' would not be highlighted, see:
https://github.com/microsoft/vscode/issues/140960

I believe this to be a regression in vscode, so I reported it:
https://github.com/microsoft/vscode/issues/168068